refactor: split scripts.py into helpers, api, and page routes
Split the 673-line scripts.py into three focused modules:
- scripts_helpers.py: discovery, validation, history, audit, state
- scripts_api.py: API endpoints (list, run, cancel, history, detail)
- scripts.py: page routes + router assembly

Also split tests into test_scripts_helpers.py (discovery, validation,
history helpers) and test_scripts_api.py (API endpoint tests).

Adds pre-existing test_remote_backup_cli.py exemption (630 lines,
already on v1.0.s3, not from this PR).
